Issue #74 Enable / disable communication (#77)
Created a new private variable: bool _enabled and two methods to change (enable() and disable()), as well as one to check the status (readEnabled()).

poll() checks the status of _enable, just before validating the request.  If enable is not true, the function returns without processing the request.

Tested as working.